+++ pytsite/http_api/_controller.py
@@ -1,0 +1,11 @@
+"""PytSite HTTP API Base Controller
+"""
+from pytsite import router as _router
+
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+class Controller(_router.Controller):
+    pass