+++ pytsite/validation/_rule.py
@@ -258,6 +258,24 @@ class GreaterOrEqual(Greater):
             raise _error.RuleError(self._msg_id, self._msg_args)
 
 
+class Choice(Base):
+    def __init__(self, value: str = None, msg_id: str = None, msg_args: dict = None, **kwargs):
+        super().__init__(value, msg_id, msg_args)
+
+        self._options = kwargs.get('options', ())
+        self._msg_args.update({'options': str(self._options)})
+
+    def _do_validate(self):
+        """Do actual validation of the rule.
+        """
+        if self._value is None:
+            return
+
+        if self._value not in self._options:
+            self._msg_args.update({'value': str(self._value)})
+            raise _error.RuleError(self._msg_id, self._msg_args)
+
+
 class Regex(Base):
     def __init__(self, value: str = None, msg_id: str = None, msg_args: dict = None, **kwargs):
         super().__init__(value, msg_id, msg_args)